Frozen Prototype File Chunk Writing/Reading Infrastructure (#68270)
This adds all the necessary infrastructure to use the reusable, single-file cache in practice:

* Create cache file in a data directory instead of a temp directory
* Fully pre-allocate it (the existing solution would at least on Linux still do a sparse allocation)
* Manage file channel resource by ref counting
* Add minimal abstraction in place of exposing `FileChannel` to allow for adjusting the concrete paging approach under the hood in a follow-up

+++ x-pack/plugin/searchable-snapshots/src/main/java/org/elasticsearch/index/store/SearchableSnapshotDirectory.java
@@ -349,7 +349,9 @@ public class SearchableSnapshotDirectory extends BaseDirectory {
         for (BlobStoreIndexShardSnapshot.FileInfo file : files()) {
             final CacheKey cacheKey = createCacheKey(file.physicalName());
             cacheService.removeFromCache(cacheKey);
-            frozenCacheService.removeFromCache(cacheKey);
+            if (partial) {
+                frozenCacheService.removeFromCache(cacheKey);
+            }
         }
     }
